Consolidate Modal files
